Sets upperbound to greedy alg. Space updates shared. Pruning works if computer can get the updated limit. Computers do not get the update

+++ homework5/src/system/TspShared.java
@@ -15,7 +15,7 @@ public class TspShared implements Shared {
 
 	@Override
 	public boolean isOlderThan(Shared other) {
-		return ((Integer) other.get() < (Integer) this.get());
+		return ((double) other.get() < (double) this.get());
 	}
 
 }
